The Bull’s Eye game
The objective of the game is to put the bull’s eye, which is on a
slider that goes from 1 to 100, as close to a randomly chosen target
value as you can. In the screenshot above, the aim is to put the bull’s
eye at 44. Because you can’t see the current value of the slider,
you’ll have to “eyeball” it.
When you’re confident of your estimate you press the “Hit Me!” button
and a popup, also known as an alert, will tell you what your score
is:The closer to the target value you are, the more points you score.
After you dismiss the alert popup by pressing the OK button, a new
round begins with a new random.target. The game repeats until the
player presses the “Start Over” button (the curly
arrow in the bottom-left corner), which resets the score to 0.
This game probably won’t make you an instant millionaire on the App
Store, but even future millionaires have to start somewhere!